Added Matching Target SC Options
When ON, matching SC will temporarily override the Element selected for nuking to match one that can burst the SC that just happened. 
Priority is: Selected Element  can burst already then we use that, if not, then we swap to element matching the day or strongest not weak on current day.