The New SCH.lua
Now the SCH.lua is the gearsets, and the code is in the SCH_Lib.lua. So people with multiple characters will able to rename SCH.lua to Character_SCH.lua without having to use the exact same gearsets file on both.  Long story short, it's a much better way to do this.